BUG try and see what pins are in the local env
